tests/wpt/tests/element-timing/picture.html

<!DOCTYPE HTML>
<head>
<meta charset=utf-8>
<title>Element Timing: observe an image rendering an alternate picture source</title>
<style>
body {
  margin: 20px;
}
#square20 {
  width: 80px;
  height: 40px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
  async_test(function (t) {
    let beforeRender;
    let img;
    assert_implements(window.PerformanceElementTiming, "PerformanceElementTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        // The reported element is the image node, but the image is the
        // alternate source.
        const pathname = window.location.origin + '/images/green-100x50.png';
        checkElement(entry, pathname, 'my_image', 'square20', beforeRender, img);
        // As we use a smaller viewport, the intersection size is smaller
        checkRect(entry, [20, 100, 20, 60]);
        checkNaturalSize(entry, 100, 50);
      })
    );
    observer.observe({entryTypes: ['element']});
    // We add the image during onload to be sure that the observer is registered
    // in time for it to observe the element timing.
    window.onload = () => {
      // Add image of width equal to 100 and height equal to 50. Use a different
      // color to verify the painted image is the alternate one.
      picture = document.createElement('picture');
      picture.id = 'rectangle';
      picture.setAttribute('elementtiming', 'my_picture');

      source = document.createElement('source');
      source.srcset = '/images/green-100x50.png'
      source.type = 'image/png';
      picture.appendChild(source);

      img = document.createElement('img');
      img.setAttribute('elementtiming', 'my_image');
      img.src = 'resources/square20.jpg'
      img.id = 'square20';
      picture.appendChild(img);

      document.body.appendChild(picture);
      beforeRender = performance.now();
    };
  }, 'Element with a rectangular image and an alternate picture image with precedence reports the correct image.');
</script>
</head>
